feat: refactors navigation to MealScreen logic
