--- components/CategoryGridTile.props.ts
@@ -1,5 +1,0 @@
-export interface CategoryGridTileProps {
-  title: string;
-  color: string;
-  onPress?: () => void;
-}
